+++ CHANGELOG.md
@@ -1,5 +1,7 @@
+### 7.00
+
 ### 6.06
-#### TBD
+#### 2025-08-18
 * LispBM:
 	* New core extansions, optimization and much more unit testing.
 	* Many bug fixes in the reader.
@@ -11,12 +13,15 @@
 		* Super fast boot possible
 		* Much easier to use const blocks
 		* Rebuild image and const data when needed
+	* Added conf-detect-lambda-enc.
+	* Added many configuration parameters.
 * New offset calibration modes and options.
 * Automatic offset calibration support.
 * Added HFI ambiguity resolution modes using id injection.
 * Support for coordinated CAN baudrate update.
 * Overmodulation support.
 * Added PWM and PWM+ABI encoder modes.
+* Detect encoder parameters during flux linkage detection.
 
 ### 6.05
 #### 2024-08-19